+++ app/gui/src/dashboard/components/AriaComponents/Menu/Menu.tsx
@@ -19,6 +19,7 @@ import type { Placement, TestIdProps } from '../types'
 import { MenuItem } from './MenuItem'
 import { MenuTrigger } from './MenuTrigger'
 
+// eslint-disable-next-line react-refresh/only-export-components
 export const MENU_STYLES = tv({
   base: 'flex flex-col [clip-path:inset(0_0_0_0_round_.75rem)] overflow-x-hidden',
   slots: {
@@ -26,6 +27,7 @@ export const MENU_STYLES = tv({
   },
 })
 
+// eslint-disable-next-line react-refresh/only-export-components
 export const MENU_SECTION_STYLES = tv({
   base: 'flex flex-col',
   slots: {
@@ -33,6 +35,7 @@ export const MENU_SECTION_STYLES = tv({
   },
 })
 
+// eslint-disable-next-line react-refresh/only-export-components
 export const MENU_SEPARATOR_STYLES = tv({
   extend: SEPARATOR_STYLES,
   base: 'my-1.5 mx-2',