+++ app/gui/src/dashboard/components/AriaComponents/Radio/RadioGroup.tsx
@@ -30,6 +30,7 @@ export interface RadioGroupProps<
   readonly style?: React.CSSProperties
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export const RADIO_GROUP_STYLES = twv.tv({
   base: 'flex flex-col gap-0.5 items-start',
   variants: { fullWidth: { true: 'w-full' } },