+++ engine/runtime-instrument-common/src/main/java/org/enso/interpreter/instrument/job/SerializeModuleJob.java
@@ -1,9 +1,9 @@
 package org.enso.interpreter.instrument.job;
 
-import java.util.logging.Level;
 import org.enso.common.CompilationStage;
 import org.enso.interpreter.instrument.execution.RuntimeContext;
 import org.enso.pkg.QualifiedName;
+import org.slf4j.LoggerFactory;
 
 /** The job that serializes module. */
 public final class SerializeModuleJob extends BackgroundJob<Void> {
@@ -32,10 +32,8 @@ public final class SerializeModuleJob extends BackgroundJob<Void> {
                   .ifPresent(
                       module -> {
                         if (module.getCompilationStage().isBefore(CompilationStage.AFTER_CODEGEN)) {
-                          ctx.executionService()
-                              .getLogger()
-                              .log(
-                                  Level.WARNING,
+                          LoggerFactory.getLogger(SerializeModuleJob.class)
+                              .warn(
                                   "Attempt to serialize the module [{}] at stage [{}].",
                                   new Object[] {module.getName(), module.getCompilationStage()});
                           return;