+++ lib/scala/pkg/src/main/scala/org/enso/pkg/Package.scala
@@ -294,6 +294,7 @@ class PackageManager[F](implicit val fileSystem: FileSystem[F]) {
     * @param edition the edition to use for the project; if not specified, it
     *                will not specify any, meaning that the current default one
     *                will be used
+    * @param jvm should JVM mode be set for the package
     * @return a package object representing the newly created package.
     */
   def create(
@@ -307,7 +308,8 @@ class PackageManager[F](implicit val fileSystem: FileSystem[F]) {
     authors: List[Contact]                   = List(),
     maintainers: List[Contact]               = List(),
     license: String                          = "",
-    componentGroups: Option[ComponentGroups] = None
+    componentGroups: Option[ComponentGroups] = None,
+    jvm: Option[Boolean]                     = None
   ): Package[F] = {
     val config = Config(
       name                 = name,
@@ -319,7 +321,8 @@ class PackageManager[F](implicit val fileSystem: FileSystem[F]) {
       edition              = edition,
       preferLocalLibraries = true,
       maintainers          = maintainers,
-      componentGroups      = componentGroups
+      componentGroups      = componentGroups,
+      jvm                  = jvm
     )
     create(root, config, template)
   }